# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Brewline API
  version: 0.1.0
  description: >-
    REST + real-time контракт Brewline.

    ФАЙЛ СГЕНЕРИРОВАН из zod-схем (src/contract/schemas.ts, ADR 0010) — править руками бессмысленно,
    правки затрёт `npm run gen:spec`.

    Типы для фронтов собираются отсюда в пакет @brewline/api-types (ADR 0009).
servers:
  - url: http://localhost:3000
    description: Локальная разработка
components:
  schemas:
    Order:
      type: object
      properties:
        orderId:
          type: string
          description: Идентификатор заказа.
        number:
          type: integer
          description: Отображаемый номер заказа.
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
          description: Состав заказа.
        source:
          $ref: "#/components/schemas/OrderSource"
        status:
          $ref: "#/components/schemas/OrderStatus"
        createdAt:
          type: string
          format: date-time
          description: Время поступления (ISO 8601). Определяет порядок в очереди.
      required:
        - orderId
        - number
        - items
        - source
        - status
        - createdAt
      description: Заказ в очереди.
    OrderItem:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          description: Название напитка.
        quantity:
          type: integer
          minimum: 1
          description: Количество порций.
      required:
        - name
        - quantity
      description: Позиция заказа.
    OrderSource:
      type: string
      enum:
        - cashier
        - web
      description: Источник заказа.
    OrderStatus:
      type: string
      enum:
        - new
        - preparing
        - ready
      description: Статус заказа. Меняется только вперёд (new → preparing → ready).
    UpdateOrderStatusRequest:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/OrderStatus"
      required:
        - status
...
paths:
...
  /orders/{orderId}/status:
    patch:
      operationId: updateOrderStatus
      summary: Сменить статус заказа
      description: Перевод статуса вперёд по автомату. Недопустимый переход отклоняется (409).
      parameters:
        - schema:
            type: string
            description: Идентификатор заказа.
          required: true
          description: Идентификатор заказа.
          name: orderId
          in: path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Обновлённый заказ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Тело запроса не соответствует контракту
        "401":
          description: Требуется действующая сессия бариста
        "404":
          description: Заказ не найден
        "409":
          description: Недопустимый переход статуса
...